🐛(howard) fix context fetching of realisation certificate
Data used to generate templates of realisation certificates are not all
tracked from the context query. In the last commit, a test of this
certificate was made through a template debug view. Comming fixes
provide a correct context for the realisation certificate and context
data not coming from context query have been all hard coded temporarily.