👷(circleci) package howard module
This repository contains sources for linked but distinct packages that
needs to be distributed separately.

We now build the two packages systematically and only new package
versions are published to pypi upon a new git tag starting with a "v",
e.g. v1.0.0 for marion and v0.1.2-howard for howard. This convention
will be documented later in the developer guide.