♻️(project) move marion from a Django project to an app
As a generic application, Marion can be installed in any existing Django
project as a library or Django application to generate PDF documents.
We've moved the base Django project to a sandbox used for development
(or integration example), and changed the `marion.documents` application
to a single distributed application: `marion`.