🔒(project) add system package upgrades in docker image builds
In Docker images, the upgrades of system packages is not run when
building. It represents a security failure, as we let failures of
previous packages versions in the images. Through system package
upgrade command executing when building images, we ensure that packages
are up to date with their latest version and security failures have been
removed.