⬇️(project) downgrade Django dependencies
The last update from renovate-bot has upgraded Django to the 4.0
version. However, pytz is no longer a Django dependency and it will
move to zoneinfo package. For now, some packages are still using pytz
and tests are failing in our tests. We have decided to keep the previous
Django version and wait for Django packages to be updated to upgrade to
Django 4.0.