🐛(marion) use static storage class to open static files
`marion.utils.static_file_fetcher` method used previously the
built-in method `open` to open static files. But it could lead to some
issue when the configured static file storage generates extra static
files (e.g: containing hash for cache purpose). In this case, the
current logic was not able to find the related static file, but using
the `open` method of the StaticFileStorage class allows us to find those
 files.