🐛(Dockerfile) fix dependencies for layout rendering
I noticed that footers/headers did not work in Marion. Then I found
out that the latest versions of WeasyPrint do not use Cairo anymore.
If we follow the installation guide in their documentation, it also
doesn't work because they forgot to mention "pango1.0-tools" as a
dependency.

With this commit, you will be able to use all the css rules of
WeasyPrint like for example declaring the page number at the top of
each page with a css rule like:
@page {
    @top-right{
        content: "Page " counter(page) " of " counter(pages);
    }
}